+++ libexec/scoop-config.ps1
@@ -26,6 +26,9 @@
 # Settings
 # --------
 #
+# debug: $true|$false
+#   Additional output will be shown to identify possible source of problems.
+#
 # proxy: [username:password@]host:port
 #   By default, Scoop will use the proxy settings from Internet Options, but with anonymous authentication.
 #
@@ -37,6 +40,9 @@
 # core.useragent: Shovel/1.0 (+https://shovel.ash258.com) PowerShell/7.2 (Windows NT 10.0; Win64; x64)
 #   Allows to configure custom useragent string for all requests.
 #   Default is automatically composed based on the current PowerShell version and OS.
+#
+# commands.cat.defaultFormat: yml|json
+#   Allows to configure preffered format of the manifest representation when using the `cat` command.
 #
 # default-architecture: 64bit|32bit|arm64
 #   Allows to configure preferred architecture for application installation.
@@ -55,9 +61,6 @@
 # NO_JUNCTIONS: $true|$false
 #   The 'current' version alias will not be used.
 #   Shims, shortcuts and environment variables will point to specific version instead.
-#
-# debug: $true|$false
-#   Additional output will be shown to identify possible source of problems.
 #
 # SCOOP_REPO: http://github.com/ScoopInstaller/Scoop
 #   Git repository containining scoop source code.